evm: move EVMOpts to types file
