+++ packages/evm/src/types.ts
@@ -5,8 +5,9 @@ import type { InterpreterStep } from './interpreter.js'
 import type { Message } from './message.js'
 import type { AsyncDynamicGasHandler, SyncDynamicGasHandler } from './opcodes/gas.js'
 import type { OpHandler } from './opcodes/index.js'
+import type { CustomPrecompile } from './precompiles/index.js'
 import type { PrecompileFunc } from './precompiles/types.js'
-import type { EVMStateManagerInterface } from '@ethereumjs/common'
+import type { Common, EVMStateManagerInterface } from '@ethereumjs/common'
 import type { Account, Address, AsyncEventEmitter } from '@ethereumjs/util'
 
 export type DeleteOpcode = {
@@ -154,6 +155,97 @@ export interface EVMInterface {
   precompiles: Map<string, PrecompileFunc>
   runCall(opts: EVMRunCallOpts): Promise<EVMResult>
   events?: AsyncEventEmitter<EVMEvents>
+}
+
+/**
+ * Options for instantiating a {@link EVM}.
+ */
+export interface EVMOpts {
+  /**
+   * Use a {@link Common} instance for EVM instantiation.
+   *
+   * ### Supported EIPs
+   *
+   * - [EIP-1153](https://eips.ethereum.org/EIPS/eip-1153) - Transient Storage Opcodes (`experimental`)
+   * - [EIP-1559](https://eips.ethereum.org/EIPS/eip-1559) - EIP-1559 Fee Market
+   * - [EIP-2315](https://eips.ethereum.org/EIPS/eip-2315) - VM simple subroutines (`experimental`)
+   * - [EIP-2565](https://eips.ethereum.org/EIPS/eip-2565) - ModExp Gas Cost
+   * - [EIP-2718](https://eips.ethereum.org/EIPS/eip-2718) - Typed Transactions
+   * - [EIP-2929](https://eips.ethereum.org/EIPS/eip-2929) - Gas cost increases for state access opcodes
+   * - [EIP-2930](https://eips.ethereum.org/EIPS/eip-2930) - Access List Transaction Type
+   * - [EIP-3198](https://eips.ethereum.org/EIPS/eip-3198) - BASEFEE opcode
+   * - [EIP-3529](https://eips.ethereum.org/EIPS/eip-3529) - Reduction in refunds
+   * - [EIP-3540](https://eips.ethereum.org/EIPS/eip-3541) - EVM Object Format (EOF) v1 (`experimental`)
+   * - [EIP-3541](https://eips.ethereum.org/EIPS/eip-3541) - Reject new contracts starting with the 0xEF byte
+   *   [EIP-3651](https://eips.ethereum.org/EIPS/eip-3651) - Warm COINBASE (`experimental`)
+   * - [EIP-3670](https://eips.ethereum.org/EIPS/eip-3670) - EOF - Code Validation (`experimental`)
+   * - [EIP-3855](https://eips.ethereum.org/EIPS/eip-3855) - PUSH0 instruction
+   * - [EIP-3860](https://eips.ethereum.org/EIPS/eip-3860) - Limit and meter initcode
+   * - [EIP-4399](https://eips.ethereum.org/EIPS/eip-4399) - Supplant DIFFICULTY opcode with PREVRANDAO (Merge)
+   * - [EIP-4788](https://eips.ethereum.org/EIPS/eip-4788) - Beacon block root in the EVM (`experimental`)
+   * - [EIP-4844](https://eips.ethereum.org/EIPS/eip-4844) - Shard Blob Transactions (`experimental`)
+   * - [EIP-4895](https://eips.ethereum.org/EIPS/eip-4895) - Beacon chain push withdrawals as operations
+   * - [EIP-5133](https://eips.ethereum.org/EIPS/eip-5133) - Delaying Difficulty Bomb to mid-September 2022
+   * - [EIP-5656](https://eips.ethereum.org/EIPS/eip-5656) - MCOPY - Memory copying instruction (`experimental`)
+   * *Annotations:*
+   *
+   * - `experimental`: behaviour can change on patch versions
+   */
+  common?: Common
+
+  /**
+   * Allows unlimited contract sizes while debugging. By setting this to `true`, the check for
+   * contract size limit of 24KB (see [EIP-170](https://git.io/vxZkK)) is bypassed.
+   *
+   * Default: `false` [ONLY set to `true` during debugging]
+   */
+  allowUnlimitedContractSize?: boolean
+
+  /**
+   * Allows unlimited contract code-size init while debugging. This (partially) disables EIP-3860.
+   * Gas cost for initcode size analysis will still be charged. Use with caution.
+   */
+  allowUnlimitedInitCodeSize?: boolean
+
+  /**
+   * Override or add custom opcodes to the EVM instruction set
+   * These custom opcodes are EIP-agnostic and are always statically added
+   * To delete an opcode, add an entry of format `{opcode: number}`. This will delete that opcode from the EVM.
+   * If this opcode is then used in the EVM, the `INVALID` opcode would instead be used.
+   * To add an opcode, add an entry of the following format:
+   * {
+   *    // The opcode number which will invoke the custom opcode logic
+   *    opcode: number
+   *    // The name of the opcode (as seen in the `step` event)
+   *    opcodeName: string
+   *    // The base fee of the opcode
+   *    baseFee: number
+   *    // If the opcode charges dynamic gas, add this here. To charge the gas, use the `i` methods of the BN, to update the charged gas
+   *    gasFunction?: function(runState: RunState, gas: BN, common: Common)
+   *    // The logic of the opcode which holds the logic of changing the current state
+   *    logicFunction: function(runState: RunState)
+   * }
+   * Note: gasFunction and logicFunction can both be async or synchronous functions
+   */
+  customOpcodes?: CustomOpcode[]
+
+  /*
+   * Adds custom precompiles. This is hardfork-agnostic: these precompiles are always activated
+   * If only an address is given, the precompile is deleted
+   * If an address and a `PrecompileFunc` is given, this precompile is inserted or overridden
+   * Please ensure `PrecompileFunc` has exactly one parameter `input: PrecompileInput`
+   */
+  customPrecompiles?: CustomPrecompile[]
+
+  /*
+   * The StateManager which is used to update the trie
+   */
+  stateManager?: EVMStateManagerInterface
+
+  /**
+   *
+   */
+  blockchain?: Blockchain
 }
 
 /**
